Adds DataSetDecorators to @DataSet
This feature aims at allowing users of the library to use tokens in their XML datasets so that they can replace them with dynamically computed values.

For instance, I use it to replace [NOW] with the date on which the dataset is loaded.
This is extremely useful when testing date dependant services.